+++ src/main/java/racingcar/RacingCarConsoleApplication.java
@@ -1,6 +1,6 @@
 package racingcar;
 
-import racingcar.controller.RacingCarController;
+import racingcar.controller.RacingCarConsoleController;
 import racingcar.view.RacingCarView;
 import racingcar.view.RacingCarViewImpl;
 
@@ -8,7 +8,7 @@ public class RacingCarConsoleApplication {
 
     public static void main(String[] args) {
         RacingCarView racingCarView = new RacingCarViewImpl();
-        RacingCarController racingCarController = new RacingCarController(racingCarView);
-        racingCarController.start();
+        RacingCarConsoleController racingCarConsoleController = new RacingCarConsoleController(racingCarView);
+        racingCarConsoleController.start();
     }
 }
