--- src/main/java/racingcar/RacingCarController.java
@@ -1,29 +1,0 @@
-package racingcar;
-
-import java.util.List;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class RacingCarController {
-
-    private final RacingCarService racingCarService;
-
-    public RacingCarController(final RacingCarService racingCarService) {
-        this.racingCarService = racingCarService;
-    }
-
-    @GetMapping("/plays")
-    public ResponseEntity<List<HistoryResponse>> getGameHistory() {
-        return ResponseEntity.ok().body(racingCarService.getHistory());
-    }
-
-    @PostMapping("/plays")
-    public ResponseEntity<PlayResponse> play(@RequestBody final PlayRequest playRequest) {
-        PlayResponse response = racingCarService.play(playRequest);
-        return ResponseEntity.ok().body(response);
-    }
-}
