+++ src/main/java/racingcar/domain/MoveStrategy.java
@@ -1,4 +1,4 @@
-package racingcar.service;
+package racingcar.domain;
 
 public abstract class MoveStrategy {
     public abstract boolean isMovable();
