+++ src/main/java/racingcar/dto/PlayRequest.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.dto;
 
 public class PlayRequest {
