+++ src/test/java/racingcar/repository/GameRepositoryTest.java
@@ -8,7 +8,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.JdbcTemplate;
-import racingcar.service.TryCount;
+import racingcar.domain.TryCount;
 
 @JdbcTest
 class GameRepositoryTest {
